test(sqlite): use `ISODATETIME` type for `now()` output to avoid pandas flaky format inference
